requirements.txt
requests==2.32.3
tenacity==8.5.0
python-dotenv==1.0.0
cachetools==5.3.2

